requirements.txt
discord.py
google-api-python-client
mysql-connector-python
pillow
aiohttp
httplib2
oauth2client
pytest
dpytest
psutil
git+https://github.com/craftspider/spidertools.git#egg=spidertools
